fix(auth): added checks for crud in template

+++ src/services/templateService.js
@@ -3,12 +3,21 @@ const prisma = new PrismaClient();
 
 const templateService = {
     getAllTemplates: async (_, args, { userId }) => {
+        if (!userId) {
+            throw new Error('Not Authorized!');
+        }
         return prisma.template.findMany({ where: { userId } });
     },
-    getTemplate: async (_, { id }) => {
+    getTemplate: async (_, { id }, { userId }) => {
+        if (!userId) {
+            throw new Error('Not Authorized!');
+        }
         return prisma.template.findUnique({ where: { id } });
     },
     addTemplate: async (_, { title, body }, { userId }) => {
+        if (!userId) {
+            throw new Error('Not Authorized!');
+        }
         if (title.length > 32) {
             throw new Error('Title cannot exceed 32 characters');
         }
@@ -17,7 +26,13 @@ const templateService = {
         });
         return data;
     },
-    updateTemplate: async (_, { title, body, id }, { role }) => {
+    updateTemplate: async (_, { title, body, id }, { userId }) => {
+        const { userId: templateUserId } = await prisma.template.findUnique({
+            where: { id },
+        });
+        if (!userId || templateUserId !== userId) {
+            throw new Error('Not Authorized!');
+        }
         if (title.length > 32) {
             throw new Error('Title cannot exceed 32 characters');
         }
@@ -27,7 +42,13 @@ const templateService = {
         });
         return data;
     },
-    deleteTemplate: async (_, { id }) => {
+    deleteTemplate: async (_, { id }, { userId }) => {
+        const { userId: templateUserId } = await prisma.template.findUnique({
+            where: { id },
+        });
+        if (!userId || templateUserId !== userId) {
+            throw new Error('Not Authorized!');
+        }
         const data = await prisma.template.delete({ where: { id } });
         return data;
     },
